+++ app/src/main/java/com/rogergcc/techjobspotter/domain/mappers/Mapper.kt
@@ -1,4 +1,4 @@
-package com.rogergcc.techjobspotter.domain
+package com.rogergcc.techjobspotter.domain.mappers
 
 
 /**
