+++ app/src/main/java/com/rogergcc/techjobspotter/ui/splash/SplashActivity.kt
@@ -32,6 +32,7 @@ class SplashActivity : AppCompatActivity() {
             }
             finish()
         }
+
     }
 
     private fun hideSystemUi() {
